Update types of legacy tracker model to match to domain one (#245)
* `score` to Double

* `tracker_id` to Long

* `last_chapter_read` to Double

* `total_chapters` to Long

* `status` to Long

+++ app/src/main/java/eu/kanade/tachiyomi/data/track/BaseTracker.kt
@@ -40,8 +40,8 @@ abstract class BaseTracker(
         return track.score
     }
 
-    override fun indexToScore(index: Int): Float {
-        return index.toFloat()
+    override fun indexToScore(index: Int): Double {
+        return index.toDouble()
     }
 
     @CallSuper
@@ -70,24 +70,24 @@ abstract class BaseTracker(
         }
     }
 
-    override suspend fun setRemoteStatus(track: Track, status: Int) {
+    override suspend fun setRemoteStatus(track: Track, status: Long) {
         track.status = status
-        if (track.status == getCompletionStatus() && track.total_chapters != 0) {
-            track.last_chapter_read = track.total_chapters.toFloat()
+        if (track.status == getCompletionStatus() && track.total_chapters != 0L) {
+            track.last_chapter_read = track.total_chapters.toDouble()
         }
         updateRemote(track)
     }
 
     override suspend fun setRemoteLastChapterRead(track: Track, chapterNumber: Int) {
         if (
-            track.last_chapter_read == 0f &&
+            track.last_chapter_read == 0.0 &&
             track.last_chapter_read < chapterNumber &&
             track.status != getRereadingStatus()
         ) {
             track.status = getReadingStatus()
         }
-        track.last_chapter_read = chapterNumber.toFloat()
-        if (track.total_chapters != 0 && track.last_chapter_read.toInt() == track.total_chapters) {
+        track.last_chapter_read = chapterNumber.toDouble()
+        if (track.total_chapters != 0L && track.last_chapter_read.toLong() == track.total_chapters) {
             track.status = getCompletionStatus()
             track.finished_reading_date = System.currentTimeMillis()
         }
